requirements.txt
python-pptx >= 0.6.18